requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
python-dotenv==1.0.1
aiosqlite==0.19.0
aiohttp-session==2.12.0
cryptography==41.0.7
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
jinja2==3.1.2

aiogram==3.4.1
aiohttp==3.9.3
python-dotenv==1.0.1
aiosqlite==0.19.0
aiohttp-session==2.12.0
cryptography==42.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
jinja2==3.1.2
